Create password creator + guesser
Before you say anything, I know. There are so, so many of these and they can be simply done with 'print(random.randint(0, 10000)), but I made something pretty new and with a lot of comments that took some time to make. I made one of these programs that generate and guesses its way to the password in a special and quick.